requirements.txt
yfinance
pandas
pyyaml
python-dateutil
requests
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
